+++ common/include/webcface/common/internal/message/base.h
@@ -6,6 +6,9 @@
 #endif
 #include "./fmt.h"
 #include "webcface/common/val_adaptor.h"
+#include <memory>
+#include <variant>
+#include <vector>
 
 #ifdef WEBCFACE_UNITY
 #include "./pack.h"