+++ common/include/webcface/common/internal/message/value.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "./base.h"
+#include "webcface/common/num_vector.h"
 #include "webcface/common/encoding.h"
 #include <vector>
 
@@ -12,7 +13,7 @@ namespace message {
 
 struct Value : public MessageBase<MessageKind::value> {
     SharedString field;
-    std::shared_ptr<std::vector<double>> data;
+    MutableNumVector data;
     MSGPACK_DEFINE_MAP(MSGPACK_NVP("f", field), MSGPACK_NVP("d", data))
 };
 
@@ -27,10 +28,10 @@ template <>
 struct Res<Value> : public MessageBase<MessageKind::value + MessageKind::res> {
     unsigned int req_id = 0;
     SharedString sub_field;
-    std::shared_ptr<std::vector<double>> data;
+    MutableNumVector data;
     Res() = default;
     Res(unsigned int req_id, const SharedString &sub_field,
-        const std::shared_ptr<std::vector<double>> &data)
+        const MutableNumVector &data)
         : req_id(req_id), sub_field(sub_field), data(data) {}
     MSGPACK_DEFINE_MAP(MSGPACK_NVP("i", req_id), MSGPACK_NVP("f", sub_field),
                        MSGPACK_NVP("d", data))